Cleanup and organize more misplaced functions
