+++ IDHANServer/src/metadata/getMetadata.cpp
@@ -1,0 +1,16 @@
+//
+// Created by kj16609 on 11/13/25.
+//
+
+#include "fgl/defines.hpp"
+#include "metadata.hpp"
+
+namespace idhan::metadata
+{
+
+drogon::Task< MetadataInfo > getMetadata( [[maybe_unused]] const RecordID record_id, [[maybe_unused]] DbClientPtr db )
+{
+	FGL_UNIMPLEMENTED();
+}
+
+} // namespace idhan::metadata
